requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
PyYAML==6.0.1
kaggle==1.5.16